refactor: get motion from db helper
