+++ src/handlers/motions/helpers.ts
@@ -1,6 +1,7 @@
 import { BigNumber } from 'ethers';
-import { mutate } from '~amplifyClient';
-import { MotionData, MotionSide } from '~types';
+import { mutate, query } from '~amplifyClient';
+import { MotionData, MotionQuery, MotionSide } from '~types';
+import { verbose } from '~utils';
 
 export * from './motionStaked/helpers';
 
@@ -17,4 +18,44 @@ export const updateMotionInDB = async (
       motionData,
     },
   });
+};
+
+/*
+ * Motion ids are not unique between voting reputation installations. If you had a motion with id of 1
+ * in a previous voting reputation installation, you will get another one after uninstalling and reinstalling
+ * the voting reputation client. So, we only want the one created most recently, since that's the one created
+ * by the currently installed version of the voting reputation extension.
+ */
+
+export const getMotionFromDB = async (
+  colonyAddress: string,
+  motionId: BigNumber,
+): Promise<MotionQuery | undefined> => {
+  const { items: motions }: { items: MotionQuery[] } = await query(
+    'getColonyMotions',
+    {
+      colonyAddress,
+    },
+  );
+
+  const motionsWithCorrectId = motions.filter(
+    ({ motionData: { motionId: id } }) => id === motionId.toString(),
+  );
+
+  if (motionsWithCorrectId.length > 1) {
+    motionsWithCorrectId.sort(
+      (a, b) =>
+        new Date(a.createdAt).valueOf() - new Date(b.createdAt).valueOf(),
+    );
+  }
+
+  const motion = motionsWithCorrectId.pop();
+
+  if (!motion) {
+    verbose(
+      'Could not find the motion in the db. This is a bug and needs investigating.',
+    );
+  }
+
+  return motion;
 };
